requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.1.0
pkg-resources==0.0.0
pyee==7.0.4
slackclient==2.9.3
slackeventsapi==2.2.1
SQLAlchemy==1.4.6
typing-extensions==3.7.4.3
Werkzeug==1.0.1
yarl==1.6.3
